add 36c3 SIM card design
this card does not use the usual 0FF size (credit card format),
but its quarter size version.
this saves shipping cost and the rest of the card it rarely used.
because of the size restriction, there is not enough space to
include the quick instructions manual on the card itself.
instead a QR code is placed with a link to the osmocom wiki.

the card has been designed for the 36C3 but is not specific to
this event.
it's a general CCC event card which could be reused for other
events.

most importantly, this design includes the templates with exact
dimensions for all possible form factors, as described by the
manufacturer.